Updating to use PNPM and updating documentation across the board
