+++ .github/dependabot.yml
@@ -6,7 +6,10 @@
 version: 2
 updates:
   - package-ecosystem: "npm" # See documentation for possible values
-    directory: "/" # Location of package manifests
+    directories: # All workspace package manifests
+      - "/"
+      - "/packages/wtc-perspective-card"
+      - "/demo"
     schedule:
       interval: "weekly"
     labels:
