+++ .github/workflows/build.yml
@@ -14,15 +14,17 @@ jobs:
       - name: Checkout Repo
         uses: actions/checkout@v4
 
-      - name: Setup python
-        uses: actions/setup-python@v5
+      - name: Setup pnpm
+        uses: pnpm/action-setup@v4
 
       - name: Setup Node.js
         uses: actions/setup-node@v4
         with:
           node-version-file: '.nvmrc'
+          cache: 'pnpm'
 
       - name: Install Dependencies
-        run: npm i
+        run: pnpm install --frozen-lockfile
 
-      - run: npm run build
+      - name: Build all packages
+        run: pnpm --recursive build
