--- .npmignore
@@ -1,5 +1,0 @@
-src/
-build/
-demo/
-.github
-.vscode
