+++ package.json
@@ -1,42 +1,24 @@
 {
-  "name": "wtc-perspective-card",
-  "version": "3.0.0-beta.0",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/wethegit/wtc-perspective-card.git"
-  },
-  "author": "Liam Egan <liam@wethecollective.com>",
-  "license": "MIT",
-  "main": "dist/wtc-perspective-card.umd.js",
-  "module": "dist/wtc-perspective-card.js",
-  "style": "dist/wtc-perspective-card.css",
-  "exports": {
-    ".": {
-      "import": "./dist/wtc-perspective-card.js",
-      "require": "./dist/wtc-perspective-card.umd.js"
-    },
-    "./style.css": "./dist/wtc-perspective-card.css",
-    "./dist/*": "./dist/*"
-  },
-  "sideEffects": false,
+  "name": "wtc-perspective-card-workspace",
+  "private": true,
+  "packageManager": "pnpm@10.13.1",
   "scripts": {
-    "dev": "vite",
-    "build": "vite build --mode lib && sass src/wtc-perspective-card.scss dist/wtc-perspective-card.css --no-source-map --style compressed",
-    "prepublishOnly": "npm run build",
-    "build:demo": "vite build",
-    "preview": "vite preview --root demo",
-    "document": "{ cat USAGE.md; npx jsdoc2md src/perspective-card.js src/clickable-perspective-card.js; } > README.md",
-    "format": "prettier --write --ignore-unknown ./src/*.{js,scss}"
+    "dev": "pnpm --filter demo dev",
+    "build": "pnpm --filter wtc-perspective-card build",
+    "build:demo": "pnpm --filter demo build",
+    "preview": "pnpm --filter demo preview",
+    "document": "pnpm --filter wtc-perspective-card document",
+    "format": "prettier --write --ignore-unknown \"packages/wtc-perspective-card/src/**/*.{js,scss}\" \"demo/*.{js,scss}\""
   },
   "engines": {
     "node": ">=24"
   },
   "devDependencies": {
-    "autoprefixer": "^10.5.0",
-    "jsdoc-to-markdown": "^9.1.3",
-    "postcss": "^8.5.15",
-    "prettier": "~3.8.4",
-    "sass": "~1.101.0",
-    "vite": "^8.0.16"
+    "prettier": "~3.8.4"
+  },
+  "pnpm": {
+    "onlyBuiltDependencies": [
+      "@parcel/watcher"
+    ]
   }
 }
