+++ packages/wtc-perspective-card/USAGE.md
@@ -1,60 +1,9 @@
-# WTC Perspective Card
+# Usage guide
 
-wtc-perspective-card provides a way to create a fake 3d card animation.
-
-## Installation
-
-```sh
-$ npm install wtc-perspective-card
-```
-
-## Demo
-
-https://codepen.io/shubniggurath/pen/99df48ac9073736b0bbf5bd0e062a096?editors=0110
-
-## Usage
-
-Import it into your project.
-
-```javascript
-import { PerspectiveCard } from 'wtc-perspective-card'
-```
-
-Import the stylesheet with sass or use the css file.
-
-```scss
-@import '~wtc-perspective-card';
-```
-
-Add your markup.
-
-```html
-<div class="card">
-  <div class="card__transformer">
-    <div class="card__artwork card__artwork--front">
-      <img
-        src="https://s3-us-west-2.amazonaws.com/s.cdpn.io/982762/9b1b5b5-1.png"
-      />
-    </div>
-    <div class="card__artwork card__artwork--rear">
-      <img
-        src="https://s3-us-west-2.amazonaws.com/s.cdpn.io/982762/pokemon_card_backside_in_high_resolution_by_atomicmonkeytcg_dah43cy-fullview.png"
-      />
-    </div>
-    <div class="card__shine"></div>
-  </div>
-</div>
-```
-
-### Instanciating
-
-#### Vanilla JS
-
-Plain vanilla javascript with ES6 and module imports.
-
-```javascript
-const card = new PerspectiveCard(document.getElementById('card'))
-```
+Everything beyond the basics: configuration, accessibility, CSS custom
+properties, events and extension points. For installation, the required
+markup and a minimal working example, see the [README](./README.md); for the
+full class/member reference, see the [API docs](./API.md).
 
 ## Data attributes
 
@@ -67,7 +16,6 @@ Both classes read configuration from data attributes on the card element. Attrib
 | `data-ambient`   | boolean / number   | off     | Enables ambient animation. Omit or set to `"false"` for hover-only. Set to `""` or `"true"` for default phase offset (`0`). Set to an integer (e.g. `"2"`) to use a custom phase offset — useful for giving each card on a page a distinct rhythm. |
 | `data-zoom`      | integer            | `40`    | How far the card lifts toward the viewer on hover. Higher values produce a more pronounced zoom effect.                                                                                                                                            |
 | `data-intensity` | integer            | `10`    | Amplitude of the ambient tilt animation. Higher values produce a wider sweep.                                                                                                                                                                      |
-| `data-debug`     | boolean (presence) | `false` | Enables debug mode when the attribute is present (no value needed).                                                                                                                                                                                |
 
 ### ClickablePerspectiveCard
 
@@ -76,7 +24,7 @@ Both classes read configuration from data attributes on the card element. Attrib
 | `data-button-label`       | string    | `"Expand"` | Accessible label for the trigger button generated by the component. Ignored if you supply your own `button.perspective-card__button` in the markup. |
 | `data-close-button`       | `"false"` | —          | Set to `"false"` to suppress the close button injected into the modal dialog.                                                                       |
 | `data-close-button-label` | string    | `"Close"`  | Accessible label for the modal close button.                                                                                                        |
-| `data-duration`           | integer   | `1500`     | Open animation duration in milliseconds. The close animation runs at ⅔ of this value (default 1000ms).                                              |
+| `data-duration`           | integer   | `800`      | Open animation duration in milliseconds. The close animation runs at ⅔ of this value.                                                               |
 | `data-start-flipped`      | boolean (presence) | `false` | When present, the card rests back-to-camera (showing its back face) and flips over to reveal its front as it opens, flipping back on close. Requires distinct front/back artwork to be visible. |
 
 ## Accessibility
