+++ packages/wtc-perspective-card/src/index.js
@@ -1,0 +1,4 @@
+export { PerspectiveCard } from './perspective-card.js'
+// The clickable card's CSSCLASSES spreads the base card's, so it is the
+// complete class-name dictionary for both.
+export { ClickablePerspectiveCard, CSSCLASSES } from './clickable-perspective-card.js'
