--- src/index.js
@@ -1,2 +1,0 @@
-export { PerspectiveCard } from './perspective-card.js'
-export { ClickablePerspectiveCard } from './clickable-perspective-card.js'
